feat: make modal work by click on mine
