+++ src/logic/add_mines_pics.js
@@ -1,11 +1,11 @@
-import { minesData } from "../data/data_mines";
-import { createFlags } from "../look/flags";
-import { loseModal } from "../look/lose_modal";
-import { createOverlay } from "../look/overlay";
-import { createCells } from "./cells";
-import { shuffleMines } from "./shuffle_mines";
+// import { minesData } from "../data/data_mines";
+// import { createFlags } from "../look/flags";
+// import { loseModal } from "../look/lose_modal";
+// import { createOverlay } from "../look/overlay";
+// import { createCells } from "./cells";
+// import { shuffleMines } from "./shuffle_mines";
 
-export function addMinesPics(minesPics, minesArray) {
+// export function addMinesPics() {
 //   document.addEventListener(
 //     "DOMContentLoaded",
 //     function () {
@@ -27,19 +27,11 @@ export function addMinesPics(minesPics, minesArray) {
 //     false
 //   );
 
-const mines = document.getElementsByClassName(".minesweeper_new_mine");
-
-console.log(mines);
-
-  for (let i = 0; i < minesArray.length; i++) {
-    mines[i].style.backgroundImage = `url(${minesPics[i]})`;
-  }
-
-  mines.forEach((element) => {
-    const body = document.querySelector(".minesweeper_body");
-    element.addEventListener("click", () => {
-      body.insertAdjacentElement("afterbegin", createOverlay());
-      body.insertAdjacentElement("afterbegin", loseModal());
-    });
-  });
-}
+//   mines.forEach((element) => {
+//     const body = document.querySelector(".minesweeper_body");
+//     element.addEventListener("click", () => {
+//       body.insertAdjacentElement("afterbegin", createOverlay());
+//       body.insertAdjacentElement("afterbegin", loseModal());
+//     });
+//   });
+// }
